refactor!: move asynchronous array operations into traits
