+++ zarrs/src/array/array_ops.rs
@@ -8,6 +8,11 @@ use super::{
     StorageTransformerChain,
 };
 use crate::config::MetadataEraseVersion;
+#[cfg(feature = "async")]
+use zarrs_storage::{
+    AsyncReadableStorageTraits, AsyncReadableWritableStorageTraits, AsyncWritableStorageTraits,
+    MaybeSend,
+};
 use zarrs_storage::{
     ReadableStorageTraits, ReadableWritableStorageTraits, StorageError, StoreKey,
     WritableStorageTraits,
@@ -19,9 +24,21 @@ mod array_ops;
 mod array_read_ops;
 mod array_update_ops;
 mod array_write_ops;
+#[cfg(feature = "async")]
+mod async_array_read_ops;
+#[cfg(feature = "async")]
+mod async_array_update_ops;
+#[cfg(feature = "async")]
+mod async_array_write_ops;
 
 pub use array_mut_ops::ArrayMutOps;
 pub use array_ops::ArrayOps;
 pub use array_read_ops::ArrayReadOps;
 pub use array_update_ops::ArrayUpdateOps;
 pub use array_write_ops::ArrayWriteOps;
+#[cfg(feature = "async")]
+pub use async_array_read_ops::AsyncArrayReadOps;
+#[cfg(feature = "async")]
+pub use async_array_update_ops::AsyncArrayUpdateOps;
+#[cfg(feature = "async")]
+pub use async_array_write_ops::AsyncArrayWriteOps;
